+++ examples/Results.php
@@ -10,13 +10,28 @@ $params = [
     "dateTo" => "2022-06-04",
 ];
 
-$tallink = new marcosraudkett\Tallink($params);
+$tallink = new marcosraudkett\Tallink();
+
+$tallink->setParams($params);
 
 // fetch journeys
 $tallink->journeys();
 
 // set outwardSailId parameter (you can get this from journeys "sailId")
 $tallink->setParam("outwardSailId", "2195288");
+
+// you can also set multiple params using setParams
+$tallink->setParams([
+    "outwardSailId" => "2195288",
+    // ** other parameters
+]);
+
+/**
+ * You could also use setParams to change the parameters an make another journeys request,
+ * collect them in results and later display e.g. 
+ * ->results()["journeys"][0] // for first request
+ * ->results()["journeys"][1] // for second request
+ */
 
 // fetch vehicle prices (for id 2195288)
 $tallink->vehiclePrices();
@@ -30,4 +45,4 @@ print_r(
 );
 
 // ** to only get journeys
-$journeys = $tallink->results()["journeys"];
+$journeys = $tallink->results()["journeys"][0];