+++ tests/VehiclesTest.php
@@ -3,9 +3,7 @@
 require_once dirname(__DIR__) . "/src/Tallink.php";
 
 test('Fetch vehicles', function () {
-    $vehicles = (new marcosraudkett\Tallink([
-        "outwardSailId" => "2195288"
-    ]))->vehiclePrices();
+    $vehicles = (new marcosraudkett\Tallink())->setParam("outwardSailId", "2195288")->vehiclePrices();
   
     expect($vehicles[0]["carCategory"])->toBeString();
 });