+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to Tallink PHP Client will be documented in this file.
 
+## 2.1.0 2022-10-07
+
+- Minor improvements
+- Parameters are now set with setParams();
+  - Parameters can't be set during getInstance(); anymore as it doesn't make sense.
+
 ## 2.0.2 2022-06-07
 - Bug fixes
   - Error handling for missing parameters